requirements.txt
tqdm==4.64.0
pandas==1.4.2
beautifulsoup4==4.9.3
requests==2.27.1
readability-lxml==0.8.1
goose3==3.1.11
random-user-agent==1.0.1
news-please==1.5.22
random-user-agent==1.0.1
playwright==1.23.1
git+https://github.com/codelucas/newspaper
selenium==4.8.0
nltk==3.8.1
